Updated flash formatter and RPI to program SPI flash via UART. Also updated code to work with custom PCB

+++ common/common_inc/fs.h
@@ -7,9 +7,8 @@
 #define DIR_EXTENSION_LEN 3
 #define ROOT_DIR_OFFSET 0x0
 
-
-uint8_t read_file(char *filepath, uint8_t *read_buffer, uint32_t max_size, uint32_t *bytes_read);
-uint8_t write_file(char* filepath, uint8_t* file_buffer, uint32_t file_size);
-uint8_t file_size(char* filepath, uint32_t* file_size);
+int read_file(char *filepath, uint8_t *read_buffer, uint32_t max_size, uint32_t *bytes_read);
+int write_file(char* filepath, uint8_t* file_buffer, uint32_t file_size);
+int file_size(char* filepath, uint32_t* file_size);
 
 #endif
